Move pipeline collection to new class
