+++ src/main/java/com/capitalone/dashboard/collector/DefaultGitlabClient.java
@@ -3,8 +3,6 @@ package com.capitalone.dashboard.collector;
 import com.capitalone.dashboard.model.*;
 import com.capitalone.dashboard.repository.*;
 import com.capitalone.dashboard.util.Supplier;
-import com.fasterxml.jackson.databind.util.ISO8601DateFormat;
-import com.google.gson.Gson;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.bson.types.ObjectId;
@@ -34,7 +32,6 @@ import java.time.ZoneOffset;
 import java.time.format.DateTimeFormatter;
 import java.time.temporal.ChronoUnit;
 import java.util.*;
-import java.util.stream.Collectors;
 
 
 /**
@@ -44,36 +41,22 @@ import java.util.stream.Collectors;
 @Component
 public class DefaultGitlabClient implements GitlabClient {
     private static final Logger LOG = LoggerFactory.getLogger(DefaultGitlabClient.class);
-    private static final int COMMITS_PAGE_SIZE = 100;
 
 
     private final RestOperations rest;
     private final CommitRepository commitRepository;
-    private final CollectorRepository collectorRepository;
-    private final CollectorItemRepository collectorItemRepository;
-    private final PipelineRepository pipelineRepository;
-    private final ComponentRepository componentRepository;
-    private final DashboardRepository dashboardRepository;
     private final GitlabSettings settings;
 
     private static final String GITLAB_API_SUFFIX = "/api/v4/";
     private static final String GITLAB_PROJECT_API_SUFFIX = String.format("%s/%s", GITLAB_API_SUFFIX, "projects");
-    private static final String COMMITS_URL_WITH_SORT = "/repository/commits?per_page=" + COMMITS_PAGE_SIZE;
 
     @Autowired
     public DefaultGitlabClient(Supplier<RestOperations> restOperationsSupplier,
                                GitlabSettings settings,
-                               CommitRepository commitRepository,
-                               CollectorRepository collectorRepository, @Qualifier("collectorItemRepository") CollectorItemRepository collectorItemRepository,
-                               PipelineRepository pipelineRepository, ComponentRepository componentRepository, DashboardRepository dashboardRepository) {
+                               CommitRepository commitRepository) {
         this.rest = restOperationsSupplier.get();
         this.settings = settings;
         this.commitRepository = commitRepository;
-        this.collectorRepository = collectorRepository;
-        this.collectorItemRepository = collectorItemRepository;
-        this.pipelineRepository = pipelineRepository;
-        this.componentRepository = componentRepository;
-        this.dashboardRepository = dashboardRepository;
     }
 
     @Override
@@ -318,175 +301,6 @@ public class DefaultGitlabClient implements GitlabClient {
         build.setEndTime(jobsForPipeline.getLastEndTime(settings.getBuildStages()));
         build.setBuildStatus(jobsForPipeline.getBuildStatus(settings.getBuildStages()));
         return build;
-    }
-
-    /**
-     * Finds or creates a pipeline for a dashboard collectoritem
-     *
-     * @param collectorItem
-     * @return
-     */
-    protected Pipeline getOrCreatePipeline(CollectorItem collectorItem) {
-        Pipeline pipeline = pipelineRepository.findByCollectorItemId(collectorItem.getId());
-        if (pipeline == null) {
-            pipeline = new Pipeline();
-            pipeline.setCollectorItemId(collectorItem.getId());
-        }
-        return pipeline;
-    }
-
-    @Override
-    public void processPipelineCommits(List<PipelineCommit> commitsOfBuildStage, ObjectId collectorId, String gitProjectId, String instanceUrl) {
-        if (commitsOfBuildStage.size() <= 0) {
-            return;
-        }
-        List<Dashboard> allDashboardsForCommit = findAllDashboardsForCollectorId(collectorId, gitProjectId);
-        List<String> dashBoardIds = allDashboardsForCommit.stream().map(d -> d.getId().toString()).collect(Collectors.toList());
-
-        String buildEnvironment = PipelineStage.BUILD.getName();
-        List<Collector> collectorList = collectorRepository.findByCollectorType(CollectorType.Product);
-        List<CollectorItem> collectorItemList = collectorItemRepository.findByCollectorIdIn(collectorList.stream().map(BaseModel::getId).collect(Collectors.toList()));
-
-
-        for (CollectorItem collectorItem : collectorItemList) {
-            boolean dashboardId = dashBoardIds.contains(collectorItem.getOptions().get("dashboardId").toString());
-            if (dashboardId) {
-                Pipeline pipeline = getOrCreatePipeline(collectorItem);
-                Map<String, EnvironmentStage> environmentStageMap = pipeline.getEnvironmentStageMap();
-                EnvironmentStage commitStage = environmentStageMap.get(PipelineStage.COMMIT.getName());
-                if (commitStage == null || commitStage.getCommits() == null || commitStage.getCommits().isEmpty()) {
-                    LOG.error("Cannot populate pipeline commits for build since no pipeline commits for Commit stage found");
-                    LOG.error("Maybe the SCM collector has not been run?");
-                    return;
-                }
-                List<PipelineCommit> pipelineCommitsOfCommitsStage = new ArrayList<>(commitStage.getCommits());
-                pipelineCommitsOfCommitsStage.sort(Comparator.comparing(PipelineCommit::getScmCommitTimestamp).reversed());
-
-                if (environmentStageMap.get(buildEnvironment) == null) {
-                    environmentStageMap.put(buildEnvironment, new EnvironmentStage());
-                }
-
-                EnvironmentStage environmentStage = environmentStageMap.get(buildEnvironment);
-                if (environmentStage.getCommits() == null) {
-                    environmentStage.setCommits(new HashSet<>());
-                }
-
-                Set<PipelineCommit> buildStageCommits = new HashSet<>();
-                //Add all existing commits and incoming commits, removing duplicates
-                buildStageCommits.addAll(environmentStage.getCommits());
-                buildStageCommits.addAll(commitsOfBuildStage);
-                Map<String, PipelineCommit> builtCommitsBySha = buildStageCommits.stream()
-                        .collect(Collectors.toMap(x -> x.getScmRevisionNumber(), x -> x));
-
-
-                List<PipelineCommit> finalSetOfBuiltCommits = new ArrayList<>();
-                long timestamp = 0;
-                for (PipelineCommit commit :
-                        pipelineCommitsOfCommitsStage) {
-                    if (builtCommitsBySha.containsKey(commit.getScmRevisionNumber())) {
-                        PipelineCommit builtPipelineCommit = builtCommitsBySha.get(commit.getScmRevisionNumber());
-                        finalSetOfBuiltCommits.add(builtPipelineCommit);
-                        timestamp = builtPipelineCommit.getTimestamp();
-                    } else {
-                        Gson gson = new Gson();
-                        PipelineCommit pipelineCommit = gson.fromJson(gson.toJson(commit), PipelineCommit.class);
-                        pipelineCommit.setTimestamp(timestamp);
-                        finalSetOfBuiltCommits.add(pipelineCommit);
-                    }
-                }
-                finalSetOfBuiltCommits.sort(Comparator.comparing(PipelineCommit::getTimestamp).reversed());
-                environmentStage.setCommits(new LinkedHashSet<>(finalSetOfBuiltCommits));
-                pipelineRepository.save(pipeline);
-            }
-        }
-    }
-
-    private List<PipelineCommit> getIntermediateCommits(PipelineCommit lastUpdatedCommit, Commit baseCommit, String gitProjectId, String instanceUrl, long timestamp) {
-        JSONArray jsonArray = getCommitsForAllPages(lastUpdatedCommit, baseCommit, gitProjectId, instanceUrl, 1, new JSONArray());
-        List<PipelineCommit> pipelineCommits = new ArrayList<>();
-        for (Object object : jsonArray) {
-            JSONObject jsonObject = (JSONObject) object;
-            JSONArray parentIdsArray = (JSONArray) jsonObject.get("parent_ids");
-            String commitId = (String) jsonObject.get("id");
-            List<Commit> scmCommits = commitRepository.findByScmRevisionNumber(commitId);
-            Commit commit;
-            if (scmCommits != null && scmCommits.size() > 0) {
-                commit = scmCommits.get(0);
-
-            } else {
-                commit = new Commit();
-                commit.setTimestamp(System.currentTimeMillis());
-                commit.setScmUrl(baseCommit.getScmUrl());
-                commit.setScmBranch(baseCommit.getScmBranch());
-                commit.setScmRevisionNumber((String) jsonObject.get("id"));
-                commit.setScmAuthor((String) jsonObject.get("author_name"));
-                commit.setScmCommitLog((String) jsonObject.get("message"));
-                long commitTimestamp = new DateTime(jsonObject.get("committed_date")).getMillis();
-                commit.setScmCommitTimestamp(commitTimestamp);
-                ArrayList<String> parentRevisions = new ArrayList<>();
-                ((JSONArray) jsonObject.get("parent_ids")).forEach(parentId -> {
-                    parentRevisions.add((String) parentId);
-
-                });
-                commit.setScmParentRevisionNumbers(parentRevisions);
-                commit.setNumberOfChanges(1);
-                commit.setType(parentRevisions.size() > 1 ? CommitType.Merge : CommitType.New);
-
-            }
-            pipelineCommits.add(new PipelineCommit(commit, timestamp));
-
-        }
-        return pipelineCommits;
-
-    }
-
-
-    private JSONArray getCommitsForAllPages(PipelineCommit lastUpdatedCommit, Commit currentCommit, String gitProjectId, String instanceUrl, int pageNum, JSONArray jsonArray) {
-        String range = "&ref_name=master&page=" + pageNum + "&since=" + new ISO8601DateFormat().format(lastUpdatedCommit.getScmCommitTimestamp())
-                 +"&until=" + new ISO8601DateFormat().format(currentCommit.getScmCommitTimestamp());
-        String commitsUrl = gitProjectId + COMMITS_URL_WITH_SORT + range; //TODO: PAGINATION
-        final String apiKey = settings.getProjectKey(gitProjectId);
-        ResponseEntity<String> commitsResponse = makeRestCall(instanceUrl + GITLAB_PROJECT_API_SUFFIX + "/" + commitsUrl, true, apiKey);
-        JSONArray commitsInCurrentStage = paresAsArray(commitsResponse);
-        jsonArray.addAll(commitsInCurrentStage);
-
-        if (commitsResponse.getHeaders() != null && !commitsResponse.getHeaders().get("X-Next-Page").get(0).equals("")) {
-            getCommitsForAllPages(lastUpdatedCommit, currentCommit, gitProjectId, instanceUrl, pageNum + 1, jsonArray);
-
-        }
-        return paresAsArray(commitsResponse);
-    }
-
-
-    private JSONArray paresAsArray(ResponseEntity<String> response) {
-        if (response == null)
-            return new JSONArray();
-        try {
-            return (JSONArray) new JSONParser().parse(response.getBody());
-        } catch (ParseException pe) {
-            LOG.debug(response.getBody());
-            LOG.error(pe.getMessage());
-        }
-        return new JSONArray();
-    }
-
-    private List<Dashboard> findAllDashboardsForCollectorId(ObjectId collectorId, String gitProjectId) {
-        List<CollectorItem> collectorItems = collectorItemRepository.findByCollectorIdIn(Collections.singletonList(collectorId));
-        if (collectorItems == null || collectorItems.size() == 0) {
-            return Collections.emptyList();
-        }
-
-        Optional<CollectorItem> collectorItemOptional =
-                collectorItems.stream().filter(item ->
-                        gitProjectId.equals(item.getOptions().get("projectId"))).findFirst();
-        if (!collectorItemOptional.isPresent()) {
-            return Collections.emptyList();
-        }
-        CollectorItem collectorItem = collectorItemOptional.get();
-        List<com.capitalone.dashboard.model.Component> components = componentRepository
-                .findByBuildCollectorItemId(collectorItem.getId());
-        List<ObjectId> componentIds = components.stream().map(BaseModel::getId).collect(Collectors.toList());
-        return dashboardRepository.findByApplicationComponentIdsIn(componentIds);
     }
 
     private PipelineJobs getJobsForPipeline(String buildUrl, String apiKey) {
