+++ src/main/java/com/capitalone/dashboard/collector/GitlabCollectorTask.java
@@ -43,6 +43,7 @@ public class GitlabCollectorTask extends CollectorTask<GitlabCollector> {
     private final GitlabSettings gitlabSettings;
     private final ComponentRepository dbComponentRepository;
     private final ConfigurationRepository configurationRepository;
+    private final PipelineCommitProcessor pipelineCommitProcessor;
 
     @Autowired
     public GitlabCollectorTask(TaskScheduler taskScheduler,
@@ -53,7 +54,7 @@ public class GitlabCollectorTask extends CollectorTask<GitlabCollector> {
                                GitlabClient gitlabClient,
                                GitlabSettings gitlabSettings,
                                ComponentRepository dbComponentRepository,
-                               ConfigurationRepository configurationRepository) {
+                               ConfigurationRepository configurationRepository, PipelineCommitProcessor pipelineCommitProcessor) {
         super(taskScheduler, "Gitlab-Build");
         this.gitlabCollectorRepository = gitlabCollectorRepository;
         this.gitlabJobRepository = gitlabJobRepository;
@@ -63,6 +64,7 @@ public class GitlabCollectorTask extends CollectorTask<GitlabCollector> {
         this.gitlabSettings = gitlabSettings;
         this.dbComponentRepository = dbComponentRepository;
 		this.configurationRepository = configurationRepository;
+        this.pipelineCommitProcessor = pipelineCommitProcessor;
     }
 
     @Override
@@ -266,7 +268,7 @@ public class GitlabCollectorTask extends CollectorTask<GitlabCollector> {
                     }
                 }
                 LOG.info("Processing pipeline commits...");
-                gitlabClient.processPipelineCommits(pipelineCommits, job.getCollectorId(), gitlabProjectId, job.getInstanceUrl());
+                pipelineCommitProcessor.processPipelineCommits(pipelineCommits, job.getCollectorId(), gitlabProjectId);
             }
             log("New builds", start, count);
         }
